Extract getProjectType() utility
